backend/requirements.txt
fastapi
pydantic
openai
python-dotenv
uvicorn
pytesseract
pillow
python-multipart
PyMuPDF